% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Country Population as a Plot}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country_name}{A name of a country as a string}
}
\value{
plot of year versus country population (in thousands)
}
\description{
This function will output a plot of a country's population in thousands for a timespan from
1950 to 2020. The input of the function must be inputted as a character string
with quotations around the name of the country of interest. The data is taken from
the World_Population excel file in this package.
}
\examples{
CountryPopulation("United States of America")
CountryPopulation("China")
}
