% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World population statistics per country from 1950-2020.}
\format{
a tibble with 16685 observations and 3 variables
\describe{
  \item{CountryName}{The name of the country, as a string.}
  \item{Year}{The year of observation, as a number.}
  \item{Population}{The associated population, as a number.}
}
}
\usage{
WorldPopulation
}
\description{
World population statistics per country from 1950-2020.
}
\keyword{datasets}
